refactor(mouse): split dispatch into button/wheel plugins
Move the button dispatch path (click/drag/selection/hyperlink) into
MouseButtonInputPlugin (button.rs) and the wheel path (scrollback/
axis-lock) into MouseWheelInputPlugin (wheel.rs), each registering its
own system per the "register in the defining file's plugin" rule.
mouse.rs keeps the shared hit-test/effect plumbing (MouseEffect,
trigger_mouse_effects, CellContext, cell_at_*) and MouseInputPlugin as a
thin aggregator. Child modules read the parent's private shared items via
super::; items the retained tests name are pub(super).

Each plugin's run_if is narrowed to the messages its path consumes
(buttons: MouseButtonInput|CursorMoved; wheel: MouseWheel) instead of the
shared OR-of-three; behavior-safe since a wheel-only frame leaves the
cursor stationary, so button drag synthesis is a no-op.

Also fix a pre-existing compile error in the axis-lock test: a stray
.flatten() on cap.0.iter() (a flat Vec<MouseEffect>) that broke the test
target build.
